* Applied Ay-Ars patch (with a few modifications) which fully melds ssl   into the server spec, which is now server:port:pass:nick:group:proto   where proto currently is "irc" or "irc-ssl". * Removed -ssl flags from "/window server" and /server. * Ay-Ars patch also includes a new function, $servertype() which returns   the proto field by way of extrapolation, (it is not the exact text   entered into that field).
* Entering a server spec in a /server or "/window server" that currently
  exists, but which has different additional field values normally
  alters the current spec to conform to the new spec.  This doesn't work
  with the proto field.  This is probably a bug.

I am somewhat uncomfortable with all of this.  I would like to alter
add_to_server_list to be the one true reference for decoding the server
spec and simplify the other functions to only breaking it into server,
port and "additional".  This may be too big a change currently.

+++ source/ctcp.c
@@ -995,7 +995,7 @@ char	*ctcp_quote_it (char *str, int len)
  * convenied, but the returned data may contain nulls!.  The len is modified
  * to contain the size of the data returned. 
  */
-char	*ctcp_unquote_it (char *str, int *len)
+char	*ctcp_unquote_it (char *str, size_t *len)
 {
 	char	*buffer;
 	char	*ptr;
